+++ source/common/api/posix/os_sys_calls_impl.cc
@@ -1,11 +1,11 @@
-#include "common/api/os_sys_calls_impl.h"
-
 #include <fcntl.h>
 #include <sys/stat.h>
 #include <unistd.h>
 
 #include <cerrno>
 #include <string>
+
+#include "common/api/os_sys_calls_impl.h"
 
 namespace Envoy {
 namespace Api {