+++ test/mocks/router/mocks.h
@@ -371,6 +371,7 @@ public:
 
   // Router::Decorator
   MOCK_METHOD(const std::string&, getOperation, (), (const));
+  MOCK_METHOD(bool, propagate, (), (const));
   MOCK_METHOD(void, apply, (Tracing::Span & span), (const));
 
   std::string operation_{"fake_operation"};